requirements.txt
pytest==8.3.3
hypothesis>=6.0
numpy>=1.26
scipy>=1.12
scikit-learn>=1.4
bayesian-optimization>=1.4
pyyaml>=6.0
mypy>=1.8
ray[default]>=2.9
prometheus_client>=0.20
structlog>=24.1
opentelemetry-api>=1.22
opentelemetry-sdk>=1.22
opentelemetry-exporter-otlp>=1.22
streamlit>=1.30
plotly>=5.18
gymnasium>=0.29
pypdf>=4.0
requests>=2.31
pydantic>=2.5
hmmlearn>=0.3
psutil>=5.9
pytest-mock>=3.12
pytest-cov>=4.1
